+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Changed
 - Add push FACT-Finder import on cron feed export
 
+### Fixed
+- Fix currency code is now taken from store config
+
 ## [v1.1.2] - 2019.06.28
 ### Changed
 - Upgrade Web Components to v3.4.0